created prompts.js file to house all inquirer prompts
